Gameplay Tweaks
Adjusted it so that New Card bonus can be used up until the last
millisecond (instead of disabling at ‘0’ — when the game still gives
you 500 more ms).

Updates to random function calls to account for update to
nrgFunctions::GetRandom(). (no longer need to pass “true” or “false” to
indicate we’re excluding a number, just pass the excluded number into
the function as last argument.